# nodejs-server-server-generated/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Whats App API"
  version: "1.0.0"
  title: "Whats App API"
  contact:
    name: "Irene-Maria Tabakis"
    email: "irma.tabakis@gmail.com"
host: "virtserver.swaggerhub.com"
basePath: "/asterakia/WhatsApp/1.0.0"
...
schemes:
- "https"
paths:
...
  /privateChats:
    post:
      tags:
      - "privatechat"
      summary: "Create a private chat between two users"
      operationId: "privateChatsPOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "userIds"
        in: "formData"
        description: "Only two userIds can be sent for this endpoint due to the fact\
          \ that you are creating a private chat"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "Created a private chat. The userIds are now being sent to\
            \ the socket functions to check if the user's are online. The variable\
            \ name of the response is chat"
          schema:
            $ref: "#/definitions/Chat"
        400:
          description: "Two userIds must be sent to the server for this endpoint to\
            \ be reach"
          schema:
            type: "string"
            example: "Two userIds must be sent to the server"
      security:
      - Authorization: []
      - Socket: []
      x-swagger-router-controller: "Privatechat"
  /privateChats/{chatId}:
    delete:
      tags:
      - "privatechat"
      summary: "Delete a private chat"
      operationId: "privateChatsChatIdDELETE"
      parameters:
      - name: "chatId"
        in: "path"
        description: "The ID of a specific chat"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "The chat was deleted"
          schema:
            $ref: "#/definitions/Chat"
      security:
      - Authorization: []
      x-swagger-router-controller: "Privatechat"
  /privateChats/{chatId}/favorite:
    put:
      tags:
      - "privatechat"
      summary: "Favorite a specific private chat"
      operationId: "privateChatsChatIdFavoritePUT"
      parameters:
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_5"
      responses:
        200:
          description: "Returns the specific chat chat which is now one of the user's\
            \ favorite chats. The variable name of the response is chat"
          schema:
            $ref: "#/definitions/Chat"
      security:
      - Authorization: []
      x-swagger-router-controller: "Privatechat"
  /privateChats/{chatId}/user/{userId}/messages:
    get:
      tags:
      - "privatechat"
      summary: "Get 40 messages belonging to a private chat."
      operationId: "privateChatsChatIdUserUserIdMessagesGET"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "chatId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      - name: "messageId"
        in: "query"
        description: "Include this query parameter when you would like to get the\
          \ next 40 messages following a specific message. If this query parameter\
          \ is not included, you will get the 40 newest messages in order"
        required: false
        type: "string"
      responses:
        200:
          description: "Returns all the messages associated with a specific private\
            \ chat. The messages will be divided into two sections -- unread & read.\
            \ In each section, the messages will be ordered from newest to oldest.\
            \ The variable name of the response is messages. allMessages will be ordered\
            \ from oldest to newest (for you Sung!). The client can figure out which\
            \ messages are unread by comparing the createdAt attribute with the last_time_user_entered\
            \ variable. This route will trigger the \"read-receipts\" socket function\
            \ for the specific user querying for message history"
          schema:
            $ref: "#/definitions/inline_response_200_14"
      security:
      - Authorization: []
      x-swagger-router-controller: "Privatechat"
...
  /messages/string:
    post:
      tags:
      - "message"
      summary: "Create a message with no image attached to the message"
      operationId: "messagesStringPOST"
      parameters:
      - name: "post"
        in: "query"
        description: " "
        required: false
        type: "boolean"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_8"
      responses:
        200:
          description: "On success, the messageInfo object will be sent to the socket\
            \ functions. The variable name of the response is messageInfo"
          schema:
            $ref: "#/definitions/Message"
        400:
          description: "This error can occur if the chatId or userID does not exist,\
            \ if the user is not a participant in the chat, or if the two users have\
            \ a relationship of blocked."
      security:
      - Authorization: []
      - Socket: []
      x-swagger-router-controller: "Message"
  /messages/media:
    post:
      tags:
      - "message"
      summary: "Send a message with a file attatched"
      operationId: "messagesMediaPOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "file"
        in: "formData"
        description: " "
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "message"
        in: "formData"
        required: false
        type: "string"
        minimum: 1
      - name: "userId"
        in: "formData"
        required: true
        type: "string"
      - name: "chatId"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns the messageInfo object. This object will be sent to\
            \ the socket functions for further processing. The variable name of the\
            \ response is messageInfo"
          schema:
            $ref: "#/definitions/Message"
        400:
          description: "This error can occur if the chatId or userId does not exist,\
            \ if the user is not a participant in the chat, or if the two users have\
            \ a relationship of blocked"
      security:
      - Authorization: []
      x-swagger-router-controller: "Message"
  /messages/file:
    post:
      tags:
      - "message"
      summary: "Send a message with a file or files"
      operationId: "messagesFilePOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "file"
        in: "formData"
        description: "Even if only one file is being, must be sent in an array. This\
          \ format allows the server to easily handle when multiple files are being\
          \ sent"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "message"
        in: "formData"
        required: false
        type: "string"
        minimum: 1
      - name: "userId"
        in: "formData"
        required: true
        type: "string"
      - name: "chatId"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "Returns the messageInfo object. This object will be sent to\
            \ the socket functions for further processing. The variable name of the\
            \ response is messageInfo"
          schema:
            $ref: "#/definitions/Message"
        400:
          description: "This error can occur if the chatId or userId does not exist,\
            \ if the user is not a participant in the chat, or if the two users have\
            \ a relationship of blocked"
      security:
      - Authorization: []
      x-swagger-router-controller: "Message"
  /messages/{messageId}:
    put:
      tags:
      - "message"
      summary: "Edit (patch) a particular message"
      operationId: "messagesMessageIdPUT"
      parameters:
      - name: "messageId"
        in: "path"
        description: "The ID of the message one wishes to edit (patch)"
        required: true
        type: "string"
        minimum: 1
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_9"
      responses:
        200:
          description: "Will return the updated message on success. The variable name\
            \ of the response is translatedMessages"
          schema:
            $ref: "#/definitions/Message"
      security:
      - Authorization: []
      x-swagger-router-controller: "Message"
  /messages/{messageId}/user/{userId}:
    delete:
      tags:
      - "message"
      summary: "delete a message"
      operationId: "messagesMessageIdUserUserIdDELETE"
      parameters:
      - name: "messageId"
        in: "path"
        description: "The ID of the message one wishes to edit (patch)"
        required: true
        type: "string"
        minimum: 1
      - name: "userId"
        in: "path"
        required: true
        type: "string"
        minimum: 1
      responses:
        200:
          description: "The message which was deleted will be return as the response.\
            \ The variable name of the response is message. The response message,\
            \ will indicate if the message was favorited or unfavorited"
          schema:
            $ref: "#/definitions/inline_response_200_18"
      security:
      - Authorization: []
      x-swagger-router-controller: "Message"
...
  /relationships:
    post:
      tags:
      - "relationship"
      summary: "Create a relationship (blocked or friends) between two users. Even\
        \ if a relationship is being updated, use this route"
      operationId: "relationshipsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_10"
      responses:
        200:
          description: "The relationship was successfully created. The variable name\
            \ of the response is relationship"
          schema:
            $ref: "#/definitions/Relationship"
      security:
      - Authorization: []
      x-swagger-router-controller: "Relationship"
  /relationships/{userId}:
    put:
      tags:
      - "relationship"
      summary: "Change a relationship from friends to blocked or from blocked back\
        \ to friends"
      operationId: "relationshipsUserIdPUT"
      parameters:
      - name: "userId"
        in: "path"
        description: "The id of the other user"
        required: true
        type: "string"
        minimum: 1
      - in: "body"
        name: "body"
        description: "Send over the userId's (user_id_1 & user_id_2) and the status"
        required: false
        schema:
          $ref: "#/definitions/Relationship"
      responses:
        200:
          description: "The new relationship will be returned. The variable name of\
            \ the response is relationship"
          schema:
            $ref: "#/definitions/Relationship"
      security:
      - Authorization: []
      x-swagger-router-controller: "Relationship"
...
securityDefinitions:
  Authorization:
    type: "apiKey"
    name: "Authorization"
    in: "header"
  Socket:
    type: "apiKey"
    name: "socket"
    in: "header"
definitions:
...
  Relationship:
    type: "object"
    properties:
      request_sender:
        type: "string"
        description: "foreign key of one user"
      request_receiver:
        type: "string"
        description: "foreign key of another user"
      status:
        type: "string"
        description: "Either friends or blocked"
    example:
      request_sender: "request_sender"
      request_receiver: "request_receiver"
      status: "status"
...
  Chat:
    type: "object"
    properties:
      id:
        type: "number"
      groupChat:
        type: "boolean"
        description: "Set to true if the chat is a group chat"
      privateChat:
        type: "boolean"
        description: "Set tot true if the chat is a private chat"
      participantsId:
        type: "array"
        example:
        - "22"
        - "33"
        description: "An array of the user Ids participating in the chat"
        items:
          type: "integer"
      last_activity:
        type: "string"
      last_activity_in_associated_channel_event_or_groupChat:
        type: "string"
    example:
      last_activity: "last_activity"
      groupChat: true
      privateChat: true
      id: 0.80082819046101150206595775671303272247314453125
      participantsId:
      - "22"
      - "33"
      last_activity_in_associated_channel_event_or_groupChat: "last_activity_in_associated_channel_event_or_groupChat"
...
  Message:
    type: "object"
    properties:
      id:
        type: "number"
      message:
        type: "string"
      chatId:
        type: "string"
      userId:
        type: "string"
      type:
        type: "string"
        example: "media, string"
      createdAt:
        type: "string"
        description: "The date time of when the message was created (sent)"
      name:
        type: "string"
        description: "The full name of the user who sent the message"
      post:
        type: "boolean"
      user:
        $ref: "#/definitions/inline_response_200_7_message_user"
      GroupChat:
        $ref: "#/definitions/Message_GroupChat"
      mediaArray:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_7_message_mediaArray"
      files:
        type: "array"
        items:
          $ref: "#/definitions/Message_files"
    example:
      createdAt: "createdAt"
      GroupChat:
        socket_id: "socket_id"
        verified: true
        last_name: "last_name"
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
        country_code: "country_code"
        passcode_expiration: "passcode_expiration"
        phone: "phone"
        id: 6.02745618307040320615897144307382404804229736328125
        position: "{}"
        first_name: "first_name"
        email: "email"
        passcode: "passcode"
      chatId: "chatId"
      post: true
      name: "name"
      files:
      - s3_url: "s3_url"
        file_string: "file_string"
      - s3_url: "s3_url"
        file_string: "file_string"
      id: 0.80082819046101150206595775671303272247314453125
      message: "message"
      type: "media, string"
      userId: "userId"
      user:
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
      mediaArray:
      - media_file_string: "media_file_string"
        media_file: "media_file"
      - media_file_string: "media_file_string"
        media_file: "media_file"
...
  inline_response_200_7_message_user:
    properties:
      avatar:
        type: "string"
      profile_picture_string:
        type: "string"
    example:
      avatar: "avatar"
      profile_picture_string: "profile_picture_string"
  inline_response_200_7_message_mediaArray:
    properties:
      media_file_string:
        type: "string"
      media_file:
        type: "string"
        description: "The url of the link to the file"
    example:
      media_file_string: "media_file_string"
      media_file: "media_file"
...
  body_5:
    type: "object"
    properties:
      userId:
        type: "string"
      trueOrFalse:
        type: "boolean"
        description: "Set to true, if you would like to make the chat a favorite.\
          \ Set to false if you would like to unfavorite a chat"
  inline_response_200_14_messages_files:
    properties:
      s3_url:
        type: "string"
      file_string:
        type: "string"
    example:
      s3_url: "s3_url"
      file_string: "file_string"
  inline_response_200_14_messages_unread:
    properties:
      id:
        type: "number"
      message:
        type: "string"
      userId:
        type: "number"
      chatId:
        type: "number"
      mediaArray:
        type: "array"
        items:
          type: "string"
      files:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_messages_files"
      type:
        type: "string"
      createdAt:
        type: "string"
      user:
        $ref: "#/definitions/inline_response_200_7_message_user"
    example:
      createdAt: "createdAt"
      chatId: 1.46581298050294517310021547018550336360931396484375
      files:
      - s3_url: "s3_url"
        file_string: "file_string"
      - s3_url: "s3_url"
        file_string: "file_string"
      id: 0.80082819046101150206595775671303272247314453125
      message: "message"
      type: "type"
      userId: 6.02745618307040320615897144307382404804229736328125
      mediaArray:
      - "mediaArray"
      - "mediaArray"
      user:
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
  inline_response_200_14_messages_read:
    properties:
      id:
        type: "number"
      message:
        type: "string"
      userId:
        type: "number"
      chatId:
        type: "number"
      mediaArray:
        type: "array"
        items:
          type: "string"
      type:
        type: "string"
      createdAt:
        type: "string"
      user:
        $ref: "#/definitions/inline_response_200_7_message_user"
    example:
      createdAt: "createdAt"
      chatId: 2.3021358869347654518833223846741020679473876953125
      id: 5.962133916683182377482808078639209270477294921875
      message: "message"
      type: "type"
      userId: 5.63737665663332876420099637471139430999755859375
      mediaArray:
      - "mediaArray"
      - "mediaArray"
      user:
        avatar: "avatar"
        profile_picture_string: "profile_picture_string"
  inline_response_200_14_messages:
    properties:
      unread:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_messages_unread"
      read:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_messages_read"
      allMessages:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_messages_read"
      last_time_user_entered:
        type: "string"
        description: "This will be the date time of the last time a user entered a\
          \ chat."
    example:
      read:
      - createdAt: "createdAt"
        chatId: 2.3021358869347654518833223846741020679473876953125
        id: 5.962133916683182377482808078639209270477294921875
        message: "message"
        type: "type"
        userId: 5.63737665663332876420099637471139430999755859375
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
      - createdAt: "createdAt"
        chatId: 2.3021358869347654518833223846741020679473876953125
        id: 5.962133916683182377482808078639209270477294921875
        message: "message"
        type: "type"
        userId: 5.63737665663332876420099637471139430999755859375
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
      last_time_user_entered: "last_time_user_entered"
      unread:
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        files:
        - s3_url: "s3_url"
          file_string: "file_string"
        - s3_url: "s3_url"
          file_string: "file_string"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
      - createdAt: "createdAt"
        chatId: 1.46581298050294517310021547018550336360931396484375
        files:
        - s3_url: "s3_url"
          file_string: "file_string"
        - s3_url: "s3_url"
          file_string: "file_string"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "type"
        userId: 6.02745618307040320615897144307382404804229736328125
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
      allMessages:
      - createdAt: "createdAt"
        chatId: 2.3021358869347654518833223846741020679473876953125
        id: 5.962133916683182377482808078639209270477294921875
        message: "message"
        type: "type"
        userId: 5.63737665663332876420099637471139430999755859375
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
      - createdAt: "createdAt"
        chatId: 2.3021358869347654518833223846741020679473876953125
        id: 5.962133916683182377482808078639209270477294921875
        message: "message"
        type: "type"
        userId: 5.63737665663332876420099637471139430999755859375
        mediaArray:
        - "mediaArray"
        - "mediaArray"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
  inline_response_200_14:
    properties:
      messages:
        $ref: "#/definitions/inline_response_200_14_messages"
    example:
      messages:
        read:
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        last_time_user_entered: "last_time_user_entered"
        unread:
        - createdAt: "createdAt"
          chatId: 1.46581298050294517310021547018550336360931396484375
          files:
          - s3_url: "s3_url"
            file_string: "file_string"
          - s3_url: "s3_url"
            file_string: "file_string"
          id: 0.80082819046101150206595775671303272247314453125
          message: "message"
          type: "type"
          userId: 6.02745618307040320615897144307382404804229736328125
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 1.46581298050294517310021547018550336360931396484375
          files:
          - s3_url: "s3_url"
            file_string: "file_string"
          - s3_url: "s3_url"
            file_string: "file_string"
          id: 0.80082819046101150206595775671303272247314453125
          message: "message"
          type: "type"
          userId: 6.02745618307040320615897144307382404804229736328125
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        allMessages:
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
        - createdAt: "createdAt"
          chatId: 2.3021358869347654518833223846741020679473876953125
          id: 5.962133916683182377482808078639209270477294921875
          message: "message"
          type: "type"
          userId: 5.63737665663332876420099637471139430999755859375
          mediaArray:
          - "mediaArray"
          - "mediaArray"
          user:
            avatar: "avatar"
            profile_picture_string: "profile_picture_string"
...
  body_8:
    type: "object"
    properties:
      userId:
        type: "string"
      message:
        type: "string"
      chatId:
        type: "string"
  body_9:
    type: "object"
    properties:
      message:
        type: "string"
  inline_response_200_18:
    properties:
      success:
        type: "boolean"
      responseMsg:
        type: "string"
      message:
        $ref: "#/definitions/Message"
    example:
      success: true
      message:
        createdAt: "createdAt"
        GroupChat:
          socket_id: "socket_id"
          verified: true
          last_name: "last_name"
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
          country_code: "country_code"
          passcode_expiration: "passcode_expiration"
          phone: "phone"
          id: 6.02745618307040320615897144307382404804229736328125
          position: "{}"
          first_name: "first_name"
          email: "email"
          passcode: "passcode"
        chatId: "chatId"
        post: true
        name: "name"
        files:
        - s3_url: "s3_url"
          file_string: "file_string"
        - s3_url: "s3_url"
          file_string: "file_string"
        id: 0.80082819046101150206595775671303272247314453125
        message: "message"
        type: "media, string"
        userId: "userId"
        user:
          avatar: "avatar"
          profile_picture_string: "profile_picture_string"
        mediaArray:
        - media_file_string: "media_file_string"
          media_file: "media_file"
        - media_file_string: "media_file_string"
          media_file: "media_file"
      responseMsg: "responseMsg"
...
  body_10:
    type: "object"
    properties:
      request_send:
        type: "string"
      request_receiver:
        type: "string"
      status:
        type: "string"
      reportBlockedUser:
        type: "string"
...
  Message_GroupChat:
    properties:
      id:
        type: "number"
        description: "Unique identifer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      phone:
        type: "string"
      country_code:
        type: "string"
      email:
        type: "string"
      profile_picture_string:
        type: "string"
      avatar:
        type: "string"
      position:
        type: "object"
        properties: {}
      verified:
        type: "boolean"
      passcode:
        type: "string"
      passcode_expiration:
        type: "string"
      socket_id:
        type: "string"
    example:
      socket_id: "socket_id"
      verified: true
      last_name: "last_name"
      avatar: "avatar"
      profile_picture_string: "profile_picture_string"
      country_code: "country_code"
      passcode_expiration: "passcode_expiration"
      phone: "phone"
      id: 6.02745618307040320615897144307382404804229736328125
      position: "{}"
      first_name: "first_name"
      email: "email"
      passcode: "passcode"
  Message_files:
    properties:
      file_string:
        type: "string"
      s3_url:
        type: "string"
    example:
      s3_url: "s3_url"
      file_string: "file_string"